# src/main/java/com/etdev/songs/api/Completion.java
package com.etdev.songs.api;

public class Completion {

    private String completion;

    public Completion(String completion) {
        this.completion = completion;
    }

    public String getCompletion() {
        return completion;
    }
}